pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-task-tracker"
version = "0.1.0"
description = "Secure Task Tracker backend built with FastAPI"
authors = [ { name = "Your Name", email = "you@example.com" } ]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["fastapi", "tasks", "jwt", "sqlalchemy"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Framework :: FastAPI",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers"
]
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "SQLAlchemy>=2.0.30",
  "alembic>=1.13.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-dotenv>=1.0.1",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "psycopg2-binary>=2.9.9",
  "python-multipart>=0.0.9",
  "email-validator>=2.1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "black>=24.4.2",
  "isort>=5.13.2",
  "ruff>=0.4.0"
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

